Add 1 customizable & specialized face to help testing
The colors of the faces are customizable.
I will create several other faces to help check the implementation of the
mode.  Later once the code stabilizes I will adjust the default colors and
attributes of the faces.

In the mean time, users can customize their copies of the faces to what they
prefer.