seed7-beg-of-defun can go to previous when on the beginning of one.
Add docstring.

Note that shift selection will not work on terminal mode because the M-C-a and
M-C-e key bindings do not allow recognizing the state of the Shift key with a
Control modifier pressed in Emacs running under a terminal. It will only work
on a GUI version of Emacs.  Another key binding is required to make it work in
terminal, such as the ones PEL uses.