Navigation verbosity now controlled by seed7-verbose-navigation
By default the seed7-verbose-navigation user-option is t, activating
verbosity in the navigation commands: on successful movement the command
prints the name of the new function or procedure found.  The message also
includes whether the mark was set or not.

When seed7-verbose-navigation user-option is nil, no message is printed.