Fix beg/enf-of-defun-function protocol
The values previously used for `beginning-of-defun-function' and
`end-of-defun-function' where the interactive functions `seed7-beg-of-defun'
and `seed7-end-of-defun' respectively.  Those interactive functions do no have
the expected arguments and do not return a non-nil/nil value depending of the
outcome.

The result was that under some conditions (ie, when comments were right before
a Seed7 function, procedure or block), code using `beg-of-defun' or
`end-of-defun' like iedit, did not behave properly.

The solution is to provided wrapper functions that behave as expected by the
protocol and use those as beginning-of-defun-function and
end-of-defun-function.  The new wrappers are:

- `seed7--beg-of-defun-simple'
- `seed7--end-of-defun-simple'

These are meant to be used indirectly for this protocol, not directly by the
user (therefore their name with the double dash).