Handle array types. Fix navigation and indentation.
The seed7-procfunc-regexp regular expression is now able to detect
functions that return a simple array type. Something like 'array integer'.
The regexp allows 2 identifier in the type.  If more is used or if some extra
non-identifiers characters are present the regexp won't work and it will have
to be modified.  But for now, I'll leave it like it is and continue testing
the mode.