Add commands that insert indented boiler code for blocks.
The commands do not yet have key bindings but are Available in the top menu
and can be invoked with M-x.

Invoke them on an empty line.  Their content will be automatically
auto-indented and the cursor is left where text must first be written.

The set is not complete yet.  The implementation is a preliminary prototype.
Please report any problem.