Add code auto complete for case, for, if, repeat & while statements.
To use: type the keyword at the beginning of the line and then hit the TAB
key.  It will write the boiler plate code for the beginning and end of the
statement block, leaving point (the cursor) at the next location to fill.

There are other blocks this could support.  I'll look into those next.