Fix block end/start navigation for functions with nested func/proc
This fixes one aspect of issue 22: https://github.com/pierre-rouleau/seed7-mode/issues/22

With this change, seed7-to-block-forward and seed7-to-block-backward move to
the corresponding end if prompt is at the end or start line.

The seed7-beg-of-defun, seed7-end-of-defun and seed7-beg-of-next-defun
still do not handle this type of code.