Early version of seed7-xref-goto.
This is an early version of the command that read the word/symbol at point
then open the file where that is defined/declared.
It works for everything defined in the file, including the operators.
It's early because it does not yet provide a way to go back to the
original location.
It provides completion, but that could be improved too.

But it works! and just uses Seed7 code to build the list and search from it.