Add proper support for xref-find-definitions and xref-go-back
Replaced the preliminary seed7-xref-goto by proper xref
framework compliant implementation of:

- xref-find-definition , bound to M-.
- xref-go-back, bound to M-,

There's still more to do: when more than 1 candidate is found,
the description of each candidate is not filled.  A message
stating that no documentation is found is included instead.

But aside from that everything expected from cross reference
navigation works.  You can use ant xref frontend: Helm for example.

This is a the first step in providing a full xref backend for Seed7.
More to come.