xref: add ability to search non referenced object in local file
If a Seed7 object reference is not found by the Seed7 s7xref program,
the `seed7-xref-get' now uses `seed7--find-info-about' to find the location of
the identifier if it is defined inside the same file.

This currently only works for identifiers at global scope.
The next step is to add ability to find local ones as well and distinguish
between them.

Also fixed the `seed7-current-line0-number' which can now work on any buffer
and inside narrowed areas.