xref: add ability to search non referenced object in local scope
If a Seed7 object reference is not found by the Seed7 s7xref program,
the `seed7-xref-get' now uses `seed7--find-info-about' to find the location of
the identifier if it is defined inside the same file.

It first looks inside the current block to find local definitions.
If it does not find any it looks inside the global scope.

There is still an issue when the block is a type definition.
2 issues remain there: it could find the same location as the current
position, and the regexp for types that are sub types is not handling it
properly.

The next step is to look into that.