Improve xref, navigation and arg parsing.  More to come.
This single commit contains a lot of code changes.
It is, however not complete.
I am in the process of providing a complete description of the
argument signature of functions and procedures via regular expression.
The s7i files use a larger set of constructs than the sd7 files do.
The detection of function/procedures use for navigation, indentation and used
to identify the scopes for xref extended navigation depend on that.
I'm almost there but have to integrate everything while being able to extract
the name of the function/procedure in all cases.  That is more difficult and
will require other changes that will be submitted later.

This commit includes:

- Update to the table of content.  Use it to navigate in the code.
- imenu control commands allowing dynamic modification of the way the entries
  are listed:
    - seed7-toggle-menu-callable-list, bound to C-c g c, toggles the way
      functions and procedures are listed in imenu driven lists: either list
      functions and procedures together inside a single group or separately.
    - seed7-toggle-menu-sorting, bound to C-c g s, toggles the way entries are
      listed: either in code order or in sorted order.
  Both of these have associated customizable user options:
   seed7-menu-list-functions-and-procedures-together and
   seed7-menu-list-functions-sorted.

Improved local/global scope search performed by  xref-based navigation invoked
by M-..  It uses a better identification of the blocks.

Added support for several keywords: attr, varfunc, DYNAMIC, fixLen, key, len,
of, subrange, subtype, interface.

Improved several regular expressions.

Wrote regular expressions to match 10 forms of parameters I have identified.

This is what needs to be integrated in the procedure/function matching.