Improve xref local identifier search: limit parameter area.
Limit the search area for parameters to the section of text inside the
parentheses.  Like local variables, the found location of the parameters done
by local searches moves point to the beginning of the identifier word.
The searches done by the s7xref table still moves the point to the beginning
of the line.