type_implements_interface is a normal identifier.
It's no longer rendered with seed7-predefined-functions-face.
I introduced it by mistake.
The mistake was triggered by another issue: the compiled version of s7xref
did not match the latest Seed7 code in my system.
The solution is to always use interpreted version of s7xref.sd7.
The next commit will address that.