xref-find-definition: prevent error when issued from definition
If cross-reference for a symbol is requested from the location of its
definition, and the visited file is modified, unsaved and that symbol just got
renamed, seed7-mode cross reference will not find it from the s7xref built
table and will search in the current buffer.  Since the found location is the
same line as the original request it wont find anything and will return a list
of one nil, which was not handled by `seed7--find-symbol'.  It is now handled
and a user-error is issued.