xref-find-definition: issue error when issued from all definitions
... including identifiers that are part of the s7xref table.

Also added 'action' to the highlighting, even though it's not really a
keyword.  But there are several other 'defined-by-language-spec-file'
identifiers that arehighlighted already because they are important to the base
language; that's just another of those.

Also renamed several defconst symbols to better represent what they deal with:
parameter instead of arguments.