Fix seed7---inner-callables-1 regexp.
That caused seed7-to-block-backward to fail on some function, causing other
issues in those Seed7 functions such as auto-indentation and cross reference
operations.

The functions affected, for example, were the ripemd160 and keccak1600
functions from lib/msgdigest.s7i

The issue was the use of "[^\\0]" regexp used to identify all characters
including newline.  This is a non-feature that seems to not be reliable.  I
will have to remove the use of that pattern from all regexp that still use it.

I will remove it but I need to perform extra testing.

This is part of the regexp cleanup and consolidation that I need to complete.