Cleanup & optimize using line-end-position
Use line-end-position to get the position at end of line.
There is a newer function, pos-eol, available in Emacs > 29,
but since seed7-mode supports Emacs starting at 26 o5 25 (not yet cast in
stone) use the older function.

Use it where the code has to move point to get the position and save excursion
to prevent moving point.  In some case the point move is necessary, but use it
and later explicitly move point to clarify the intent and reduce risk of
breaking code later in maintenance.