Add more adapted protection against issue #31
Issue #31 is most likely caused by an incorrect or incomplete installation of
seed7-mode.

That can happen.

One scenario is that the seed7-mode package was not installed, but just the
seed7-mode.el file was opened (like Andreas Röhler seems to have done).
The added code uses the directory where the seed7-mode.el file is stored if
the seed7-mode is not in a directory located in Emacs `load-path'.

That might still not be the case, so I kept the last default of a
~/.emacs.d/utils directory.