Fix seed7-line-inside-a-block when using dummy noop injection
In some circumstances, seed7-line-inside-a-block temporarily inserts a noop
command in the buffer to compute the indentation and then removed it.
The old code did not properly cleanup after that and did not support read-only
buffer. The new code handles read-only buffers and properly cleans up the undo
history.

This function is used in several commands, including the xref commands; that
caused the xref commands to fail in read-only buffer.  It also injected some
operations in the undo list buffer.

This is now fixed.